requirements.txt
streamlit
pandas
pymongo
hyperliquid-python
